requirements.txt
pandas
numpy
scikit-learn
rapidfuzz
openai
snowflake-connector-python
streamlit
python-dotenv
